backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
amqp==5.3.1
attrs==24.2.0
beautifulsoup4==4.12.2
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.0
celery==5.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
edge-tts==6.1.19
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
langdetect==1.0.9
lxml==5.3.0
MarkupSafe==3.0.2
moviepy==2.1.1
multidict==6.1.0
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydub==0.25.1
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-pptx==0.6.21
redis==5.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.6.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webvtt-py==0.5.1
Werkzeug==3.1.3
XlsxWriter==3.2.0
yarl==1.18.3
